NewsPaper/requirements.txt
Django==5.1.1
django-allauth==0.61.1
django-filter==24.1
python-dotenv==1.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-debug-toolbar==4.3.0
django-environ==0.11.2
requests==2.31.0 